backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
groq
PyMuPDF
azure-storage-blob
chromadb
sentence-transformers
requests
beautifulsoup4
crawl4ai
httpx
tqdm
numpy
pandas